fix: 중복파일삭제, 예약 중복 삽입 수정
