+++ src/features/payment/pages/UserPayment.css
@@ -33,8 +33,6 @@
   height: 56px;
   display: flex;
   align-items: center;
-  border-bottom: 1px solid #e5e7eb;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   padding: 0 16px;
 }
 
@@ -428,4 +426,94 @@ html {
   .payment-method-item {
     padding: 12px;
   }
+}
+
+/* 카드 정보 입력 섹션 */
+.card-info-section,
+.bank-info-section,
+.kakao-info-section {
+  margin-bottom: 32px;
+}
+
+.card-inputs {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.input-group {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.input-group label {
+  font-size: 14px;
+  font-weight: 600;
+  color: #333;
+}
+
+.input-group input {
+  padding: 12px 16px;
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  font-size: 16px;
+  transition: border-color 0.2s ease;
+}
+
+.input-group input:focus {
+  outline: none;
+  border-color: #4285f4;
+  box-shadow: 0 0 0 3px rgba(66, 133, 244, 0.1);
+}
+
+.input-row {
+  display: flex;
+  gap: 16px;
+}
+
+.input-row .input-group {
+  flex: 1;
+}
+
+/* 무통장입금 정보 */
+.bank-details {
+  background-color: #f8f9fa;
+  padding: 20px;
+  border-radius: 8px;
+  border-left: 4px solid #4285f4;
+}
+
+.bank-details p {
+  margin: 8px 0;
+  font-size: 15px;
+  color: #333;
+}
+
+.bank-note {
+  color: #666 !important;
+  font-size: 13px !important;
+  font-style: italic;
+  margin-top: 16px !important;
+}
+
+/* 카카오페이 정보 */
+.kakao-details {
+  background-color: #fff8e1;
+  padding: 20px;
+  border-radius: 8px;
+  border-left: 4px solid #ffb300;
+}
+
+.kakao-details p {
+  margin: 8px 0;
+  font-size: 15px;
+  color: #333;
+}
+
+.kakao-note {
+  color: #666 !important;
+  font-size: 13px !important;
+  font-style: italic;
+  margin-top: 16px !important;
 }
