+++ src/features/payment/pages/UserPayment.jsx
@@ -8,7 +8,13 @@ import { usePaymentData } from '../../reservation/hooks/useLocalStorage';
 
 const UserPayment = () => {
   const navigate = useNavigate();
-  const [selectedPaymentMethod, setSelectedPaymentMethod] = useState('bank');
+  const [selectedPaymentMethod, setSelectedPaymentMethod] = useState('card');
+  const [cardInfo, setCardInfo] = useState({
+    number: '',
+    expiry: '',
+    cvc: '',
+    name: '',
+  });
 
   // zustand store에서 예약 데이터 가져오기
   const { reservationData, getSelectedServicesWithDetails } =
@@ -18,13 +24,41 @@ const UserPayment = () => {
   // localStorage에서 결제 데이터 가져오기
   const { paymentData: savedPaymentData } = usePaymentData();
 
-  // ⭐️ 디버깅: 결제 데이터 확인
+  // 결제 데이터가 없으면 홈으로 리다이렉트
   useEffect(() => {
-    console.log('💳 UserPayment - 결제 데이터 상태:');
-    console.log('📊 zustand reservationData:', reservationData);
-    console.log('💾 localStorage paymentData:', savedPaymentData);
-    console.log('🛍️ selectedServices:', selectedServices);
-  }, [reservationData, savedPaymentData, selectedServices]);
+    if (!savedPaymentData && reservationData.selectedServices.length === 0) {
+      alert('결제 정보가 없습니다.');
+      navigate('/');
+    }
+  }, [savedPaymentData, reservationData, navigate]);
+
+  // 카드 입력 헬퍼 함수들
+  const handleCardInputChange = (field, value) => {
+    setCardInfo((prev) => ({ ...prev, [field]: value }));
+  };
+
+  const formatCardNumber = (value) => {
+    const v = value.replace(/\s+/g, '').replace(/[^0-9]/gi, '');
+    const matches = v.match(/\d{4,16}/g);
+    const match = (matches && matches[0]) || '';
+    const parts = [];
+    for (let i = 0, len = match.length; i < len; i += 4) {
+      parts.push(match.substring(i, i + 4));
+    }
+    if (parts.length) {
+      return parts.join(' ');
+    } else {
+      return v;
+    }
+  };
+
+  const formatExpiry = (value) => {
+    const v = value.replace(/\s+/g, '').replace(/[^0-9]/gi, '');
+    if (v.length >= 2) {
+      return v.substring(0, 2) + '/' + v.substring(2, 4);
+    }
+    return v;
+  };
 
   // 결제 데이터 생성 (localStorage 데이터 우선 사용)
   const paymentData = {
@@ -39,8 +73,12 @@ const UserPayment = () => {
       serviceType:
         savedPaymentData?.serviceInfo?.type ||
         savedPaymentData?.serviceInfo?.subOptionName ||
+        reservationData.selectedSubOption?.name ||
         '청소 서비스',
       manager: '매니저 배정 예정',
+      reservationId: savedPaymentData?.reservationId,
+      status: savedPaymentData?.status || 'PENDING',
+      duration: savedPaymentData?.duration,
     },
     priceList:
       selectedServices.length > 0
@@ -60,16 +98,55 @@ const UserPayment = () => {
       (reservationData.totalPrice > 0 ? reservationData.totalPrice : 155000),
   };
 
-  const handlePayment = () => {
-    // 결제 처리 로직
-    console.log('💳 결제 처리 중...');
-    console.log('💰 결제 데이터:', paymentData);
-    console.log('📋 예약 데이터:', reservationData);
-    console.log('📄 저장된 결제 데이터:', savedPaymentData);
+  const handlePayment = async () => {
+    // 카드 정보 검증
+    if (selectedPaymentMethod === 'card') {
+      if (
+        !cardInfo.number ||
+        !cardInfo.expiry ||
+        !cardInfo.cvc ||
+        !cardInfo.name
+      ) {
+        alert('카드 정보를 모두 입력해주세요.');
+        return;
+      }
+    }
 
-    // 결제 완료 시뮬레이션
-    alert('결제가 성공적으로 완료되었습니다!');
-    navigate('/user/payment-complete');
+    try {
+      // Spring Boot 결제 API 호출을 위한 데이터 준비
+      const paymentRequestData = {
+        reservationId: paymentData.serviceInfo.reservationId,
+        paymentMethod: selectedPaymentMethod,
+        amount: paymentData.totalAmount,
+        paymentDetails:
+          selectedPaymentMethod === 'card'
+            ? {
+                cardNumber: cardInfo.number.replace(/\s/g, ''),
+                expiryDate: cardInfo.expiry,
+                cvc: cardInfo.cvc,
+                cardHolderName: cardInfo.name,
+              }
+            : null,
+      };
+
+      // TODO: 실제 Spring Boot 결제 API 연동
+      // const paymentResult = await customerAPI.processPayment(paymentRequestData);
+      console.log('결제 요청 데이터 (추후 API 연동용):', paymentRequestData);
+
+      // 임시로 결제 성공 처리
+      alert(
+        `결제가 완료되었습니다!${
+          paymentData.serviceInfo.reservationId
+            ? `\n예약번호: ${paymentData.serviceInfo.reservationId}`
+            : ''
+        }\n결제금액: ${paymentData.totalAmount.toLocaleString()}원`
+      );
+
+      navigate('/user/payment-complete');
+    } catch (error) {
+      console.error('결제 실패:', error);
+      alert('결제에 실패했습니다. 다시 시도해주세요.');
+    }
   };
 
   const handleCancel = () => {
@@ -103,21 +180,65 @@ const UserPayment = () => {
                   {paymentData.serviceInfo.serviceType}
                 </span>
               </div>
+              {paymentData.serviceInfo.duration && (
+                <div className="info-item">
+                  <span className="label">예상 소요시간</span>
+                  <span className="value">
+                    {paymentData.serviceInfo.duration}분
+                  </span>
+                </div>
+              )}
               <div className="info-item">
                 <span className="label">매니저</span>
                 <span className="value">{paymentData.serviceInfo.manager}</span>
               </div>
-              {reservationData.address && (
+              {/* 통합된 서비스 주소 표시 (위도, 경도 값 우선) */}
+              {(reservationData.address ||
+                savedPaymentData?.serviceInfo?.address) && (
                 <div className="info-item">
                   <span className="label">서비스 주소</span>
                   <span className="value">
-                    {/* 위도, 경도가 포함된 addressDetail만 표시하거나, address가 좌표 정보인 경우 처리 */}
-                    {reservationData.addressDetail &&
-                    reservationData.addressDetail.includes('위도:')
-                      ? reservationData.addressDetail
-                      : reservationData.address.includes('위도:')
-                        ? reservationData.address
-                        : `${reservationData.address} ${reservationData.addressDetail || ''}`}
+                    {(() => {
+                      // 1. reservationData에서 위도,경도 정보 확인
+                      if (
+                        reservationData.addressDetail &&
+                        reservationData.addressDetail.includes('위도:')
+                      ) {
+                        return reservationData.addressDetail;
+                      }
+                      if (
+                        reservationData.address &&
+                        reservationData.address.includes('위도:')
+                      ) {
+                        return reservationData.address;
+                      }
+
+                      // 2. savedPaymentData에서 위도,경도 정보 확인
+                      if (
+                        savedPaymentData?.serviceInfo?.address?.detail &&
+                        savedPaymentData.serviceInfo.address.detail.includes(
+                          '위도:'
+                        )
+                      ) {
+                        return savedPaymentData.serviceInfo.address.detail;
+                      }
+
+                      // 3. 그 외의 경우 일반 주소 표시
+                      if (savedPaymentData?.serviceInfo?.address) {
+                        return `${savedPaymentData.serviceInfo.address.main}${
+                          savedPaymentData.serviceInfo.address.detail
+                            ? ` ${savedPaymentData.serviceInfo.address.detail}`
+                            : ''
+                        }`;
+                      }
+
+                      // 4. 마지막으로 reservationData 주소 표시
+                      return `${reservationData.address}${
+                        reservationData.addressDetail
+                          ? ` ${reservationData.addressDetail}`
+                          : ''
+                      }`;
+                    })()}
                   </span>
                 </div>
               )}
@@ -163,6 +284,19 @@ const UserPayment = () => {
                 <input
                   type="radio"
                   name="paymentMethod"
+                  value="card"
+                  checked={selectedPaymentMethod === 'card'}
+                  onChange={(e) => setSelectedPaymentMethod(e.target.value)}
+                />
+                <div className="method-content">
+                  <i className="fas fa-credit-card method-icon"></i>
+                  <span className="method-text">신용카드/체크카드</span>
+                </div>
+              </label>
+              <label className="payment-method-item">
+                <input
+                  type="radio"
+                  name="paymentMethod"
                   value="bank"
                   checked={selectedPaymentMethod === 'bank'}
                   onChange={(e) => setSelectedPaymentMethod(e.target.value)}
@@ -176,17 +310,112 @@ const UserPayment = () => {
                 <input
                   type="radio"
                   name="paymentMethod"
-                  value="card"
-                  checked={selectedPaymentMethod === 'card'}
+                  value="kakao"
+                  checked={selectedPaymentMethod === 'kakao'}
                   onChange={(e) => setSelectedPaymentMethod(e.target.value)}
                 />
                 <div className="method-content">
-                  <i className="fas fa-credit-card method-icon"></i>
-                  <span className="method-text">신용카드</span>
+                  <i className="fas fa-comment method-icon"></i>
+                  <span className="method-text">카카오페이</span>
                 </div>
               </label>
             </div>
           </div>
+
+          {/* 카드 정보 입력 */}
+          {selectedPaymentMethod === 'card' && (
+            <div className="card-info-section">
+              <h3 className="section-title">카드 정보</h3>
+              <div className="card-inputs">
+                <div className="input-group">
+                  <label>카드번호</label>
+                  <input
+                    type="text"
+                    placeholder="0000 0000 0000 0000"
+                    value={cardInfo.number}
+                    onChange={(e) =>
+                      handleCardInputChange(
+                        'number',
+                        formatCardNumber(e.target.value)
+                      )
+                    }
+                    maxLength="19"
+                  />
+                </div>
+                <div className="input-row">
+                  <div className="input-group">
+                    <label>유효기간</label>
+                    <input
+                      type="text"
+                      placeholder="MM/YY"
+                      value={cardInfo.expiry}
+                      onChange={(e) =>
+                        handleCardInputChange(
+                          'expiry',
+                          formatExpiry(e.target.value)
+                        )
+                      }
+                      maxLength="5"
+                    />
+                  </div>
+                  <div className="input-group">
+                    <label>CVC</label>
+                    <input
+                      type="text"
+                      placeholder="000"
+                      value={cardInfo.cvc}
+                      onChange={(e) =>
+                        handleCardInputChange(
+                          'cvc',
+                          e.target.value.replace(/[^0-9]/g, '')
+                        )
+                      }
+                      maxLength="3"
+                    />
+                  </div>
+                </div>
+                <div className="input-group">
+                  <label>카드소지자명</label>
+                  <input
+                    type="text"
+                    placeholder="홍길동"
+                    value={cardInfo.name}
+                    onChange={(e) =>
+                      handleCardInputChange('name', e.target.value)
+                    }
+                  />
+                </div>
+              </div>
+            </div>
+          )}
+
+          {/* 무통장입금 정보 */}
+          {selectedPaymentMethod === 'bank' && (
+            <div className="bank-info-section">
+              <h3 className="section-title">입금 정보</h3>
+              <div className="bank-details">
+                <p>은행: 국민은행</p>
+                <p>계좌번호: 123-456-789012</p>
+                <p>예금주: (주)홈에이드</p>
+                <p className="bank-note">
+                  * 입금 후 확인까지 1-2시간 정도 소요될 수 있습니다.
+                </p>
+              </div>
+            </div>
+          )}
+
+          {/* 카카오페이 정보 */}
+          {selectedPaymentMethod === 'kakao' && (
+            <div className="kakao-info-section">
+              <h3 className="section-title">카카오페이</h3>
+              <div className="kakao-details">
+                <p>카카오페이로 간편하게 결제하세요.</p>
+                <p className="kakao-note">
+                  * 결제 버튼 클릭 시 카카오페이 앱으로 이동합니다.
+                </p>
+              </div>
+            </div>
+          )}
 
           {/* 버튼 섹션 */}
           <div className="button-section">
