+++ src/features/payment/pages/UserPaymentComplete.css
@@ -32,8 +32,6 @@
   height: 56px;
   display: flex;
   align-items: center;
-  border-bottom: 1px solid #e5e7eb;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   padding: 0 16px;
 }
